pi-camera/requirements.txt
request
netifaces